gestciv/requirement.txt
appdirs==1.4.3
asgiref==3.2.7
bootstrap3-datetime==2.4
CacheControl==0.12.6
cairocffi==1.2.0
CairoSVG==2.5.1
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cssselect2==0.4.1
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.4
django-ajax-selects==2.1.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==14.2.0
django-countries==7.0
django-excel-fixture==0.1
django-extensions==3.1.0
django-weasyprint==1.0.2
djangorestframework==3.11.0
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
jdcal==1.4.1
lockfile==0.12.2
msgpack==0.6.2
numpy==1.20.0
openpyxl==3.0.6
packaging==20.3
pandas==1.2.1
pdfkit==0.6.1
pdfminer==20191125
pep517==0.8.2
Pillow==7.1.1
progress==1.5
psycopg2-binary==2.8.5
pycparser==2.20
pycryptodome==3.9.9
pyparsing==2.4.6
PyPDF2==1.26.0
PyPDF4==1.27.0
Pyphen==0.10.0
python-dateutil==2.8.1
python-decouple==3.3
pytoml==0.1.21
pytz==2019.3
requests==2.22.0
retrying==1.3.3
simplejson==3.17.2
six==1.14.0
slate==0.3
sqlparse==0.3.1
tabula-py==2.2.0
tinycss2==1.1.0
urllib3==1.25.8
WeasyPrint==52.2
webencodings==0.5.1
